backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-extensions==3.2.3
django-cors-headers==4.7.0
django-extensions==3.2.3

celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1

pytz==2025.1

pytest==8.3.5
pytest-django==4.10.0

gunicorn==23.0.0

psycopg2==2.9.10

python-dotenv==1.1.0
debugpy==1.8.14
requests==2.32.3
redis==6.2.0
django-redis==5.4.0